moved recommendations view components from pages to components folder
