+++ client/src/components/Recommendations/RecommendationsView/RecommendationsView.tsx
@@ -232,3 +232,4 @@ export function RecommendationsView({
     </div>
   );
 }
+
