+++ client/src/components/Recommendations/RecommendationsView/index.ts
@@ -1,0 +1,4 @@
+export { RecommendationsView } from "./RecommendationsView";
+export { useRecommendations } from "./useRecommendations";
+export type { PanelMode } from "./useRecommendations";
+
